refactor season screen
