+++ app/src/main/java/com/giraffe/cineverseapp/di/viewModelModule.kt
@@ -6,6 +6,7 @@ import com.giraffe.details.screens.castDetails.CastDetailsViewModel
 import com.giraffe.details.screens.moviedetails.MovieDetailsViewModel
 import com.giraffe.details.screens.recommended.movies.RecommendedMoviesViewModel
 import com.giraffe.details.screens.recommended.series.RecommendedSeriesViewModel
+import com.giraffe.details.screens.seasons.SeasonsViewModel
 import com.giraffe.details.screens.seriesdetails.SeriesDetailsViewModel
 import com.giraffe.explore.screen.discover.DiscoverViewModel
 import com.giraffe.explore.screen.search.SearchViewModel
@@ -31,4 +32,5 @@ val viewModelModule = module {
         )
     }
     viewModelOf(::RecommendedSeriesViewModel)
+    viewModelOf(::SeasonsViewModel)
 }
