+++ presentation/details/src/main/java/com/giraffe/details/screens/seasons/SeasonsScreenState.kt
@@ -4,5 +4,5 @@ import com.giraffe.details.models.SeasonUi
 
 data class SeasonsScreenState(
     val seasons: List<SeasonUi> = emptyList(),
-    val isLoadingSeason: Boolean = true
+    val isLoading: Boolean = true
 )
