+++ presentation/details/src/main/java/com/giraffe/details/screens/seasons/SeasonsViewModel.kt
@@ -1,16 +1,21 @@
 package com.giraffe.details.screens.seasons
 
+import androidx.lifecycle.SavedStateHandle
+import androidx.navigation.toRoute
 import com.giraffe.details.base.BaseViewModel
 import com.giraffe.details.models.SeasonUi
+import com.giraffe.details.screens.seasons.screen.SeasonsRoute
 import com.giraffe.media.series.entity.Season
 import com.giraffe.media.series.usecase.GetLastSeasonsUseCase
 
 class SeasonsViewModel(
-    seriesID: Int,
-    private val getSeasons: GetLastSeasonsUseCase
+    private val getSeasons: GetLastSeasonsUseCase,
+    savedStateHandle: SavedStateHandle
 ) : BaseViewModel<SeasonsScreenState, SeasonsEffect>(
     SeasonsScreenState()
 ) {
+
+    val seriesID = savedStateHandle.toRoute<SeasonsRoute>().seriesID
 
     init {
         loadSeason(seriesID)
@@ -29,7 +34,7 @@ class SeasonsViewModel(
         updateState {
             it.copy(
                 seasons = season.map { SeasonUi.fromEntity(it) },
-                isLoadingSeason = false
+                isLoading = false
             )
         }
     }
@@ -37,7 +42,7 @@ class SeasonsViewModel(
     fun loadSeasonsError(error: Throwable) {
         updateState {
             it.copy(
-                isLoadingSeason = false,
+                isLoading = false,
             )
         }
         sendEffect(SeasonsEffect.Error(error))
